Tranferred alarms to alarm screen
